Add game
Now we can add games
update setup.sh
update RetroPie folder setting process

+++ AddGame/addGame_upload.php
@@ -7,10 +7,20 @@
   	html_generate_top("File Upload");
 
 
-$target_dir = "/var/www/html/uploads/";
-$target_name = $target_dir . basename($_FILES["gameToUpload"]["name"]);
+
+$target_dir = $_POST['gamePlatform'];
+$target_name = $target_dir . "/" . basename($_FILES["gameToUpload"]["name"]);
+
+if(file_exists($target_dir) === FALSE){
+	echo "<strong>ERROR</strong> : Invalid destination, aborting ...";
+	html_generate_bottom();
+	exit();
+}
 
 if(move_uploaded_file($_FILES["gameToUpload"]["tmp_name"], $target_name)){
+	if(chmod($target_name, 0777) === FALSE){
+		echo "<strong>Warning</strong> Can't edit game file privilege, may result in some issue.";
+	}
 	echo "Game sucessfully uploaded : return to <a href=\"/\">HOME PAGE</a>.";
 } else {
 	echo "Error while uploading game : <a href=\"addGame.php\">Try Again</a>.";
